+++ src/main/java/com/pcs/domain/part/service/PartService.java
@@ -58,6 +58,7 @@ public class PartService {
             "B",
             "C",
             "DEFECTIVE",
+            "CANCELED",
             "OUTBOUND"
     );